+++ index.html
@@ -68,10 +68,7 @@
     </div>
   </footer>
 
-
-  <!-- <script src="index.js"  type="module"></script> -->
   <script src="index.js" type="module"></script>
-  <script src="/modules/dynamic-show.js" type="module"></script>
 </body>
 
 </html>
