+++ modules/time.js
@@ -15,8 +15,8 @@ function refreshTime() {
 function initTime() {
   refreshTime();
   setInterval(refreshTime, 1000);
-}
+} 
 
-initTime();
+// initTime();
 
 export { refreshTime, initTime };
